requirements.txt
Flask
google-generativeai==0.8.4
python-dotenv
